requirements.txt
djangorestframework==3.12.4
asgiref==3.4.1
Django==4.0
sqlparse==0.4.2
tzdata==2021.5
pytz==2021.3
djangorestframework-simplejwt==4.7.2